// tests-unit/enum-union.gen.ts
export const Kind = {
  USER: 'USER',
  ADMIN: 'ADMIN',
} as const
export type Kind = (typeof Kind)[keyof typeof Kind]

export const Intent = {
  openSession: 'openSession',
  closeSession: 'closeSession',
} as const
export type Intent = (typeof Intent)[keyof typeof Intent]

export interface UserRec {
  kind: Kind
  intent: Intent
}
// ...
export const errors = {
  WebrpcEndpoint: 'WebrpcEndpoint',
  WebrpcRequestFailed: 'WebrpcRequestFailed',
  WebrpcBadRoute: 'WebrpcBadRoute',
  WebrpcBadMethod: 'WebrpcBadMethod',
  WebrpcBadRequest: 'WebrpcBadRequest',
  WebrpcBadResponse: 'WebrpcBadResponse',
  WebrpcServerPanic: 'WebrpcServerPanic',
  WebrpcInternalError: 'WebrpcInternalError',
  WebrpcClientAborted: 'WebrpcClientAborted',
  WebrpcStreamLost: 'WebrpcStreamLost',
  WebrpcStreamFinished: 'WebrpcStreamFinished',
  Unauthorized: 'Unauthorized',
  RateLimited: 'RateLimited',
} as const
export type errors = (typeof errors)[keyof typeof errors]

export const WebrpcErrorCodes = {
  WebrpcEndpoint: 0,
  WebrpcRequestFailed: -1,
  WebrpcBadRoute: -2,
  WebrpcBadMethod: -3,
  WebrpcBadRequest: -4,
  WebrpcBadResponse: -5,
  WebrpcServerPanic: -6,
  WebrpcInternalError: -7,
  WebrpcClientAborted: -8,
  WebrpcStreamLost: -9,
  WebrpcStreamFinished: -10,
  Unauthorized: 1000,
  RateLimited: 1001,
} as const
export type WebrpcErrorCodes = (typeof WebrpcErrorCodes)[keyof typeof WebrpcErrorCodes]